pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "akimous"
version = "0.4.0"
description = "An intelligent Python IDE"
authors = ["Yu-Ann Chen <red8012@gmail.com>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://akimous.com"
repository = "https://github.com/akimous/akimous"
documentation = "https://akimous.com/start"
keywords = ["ide", "editor", "akimous"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development",
    "Topic :: Text Editors :: Integrated Development Environments (IDE)"
]
packages = [
    { include = "akimous" },
    { include = "akimous_ui" },
]
include = ["akimous_ui/**/*"]
exclude = ["akimous/demo/**/*", "akimous/scripts/**/*"]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.15"
gitpython = "^3.0"
jedi = "^0.15.0"
sphinx = "^2.0"
watchdog = "^0.9.0"
logzero = "^1.5"
websockets = "^8.0"
xgboost = "^0.90"
numpydoc = "^0.9.1"
fuzzywuzzy = "^0.17.0"
python-levenshtein = "^0.12.0"
msgpack = "^0.6"
wordsegment = "^1.3"
boltons = "^19.0"
pylint = "^2.1"
pyflakes = "^2.0"
yapf = "^0.29.0"
pathspec = "^0.6.0"
cachetools = "^3.0"
appdirs = "^1.4"
ptyprocess = "^0.6.0"
jupyter_client = "^5.2"
ipykernel = "^5.1"
toml = "^0.10.0" # required by isort
Send2Trash = "^1.5"

[tool.poetry.dev-dependencies]
pandas = "^0.25.0"
scipy = "^1.1"
pytest = "^5.0"
colorama = "^0.4"
tqdm = "^4.26"
# matplotlib = "^3.0" # for test purpose only, not really used
sympy = "^1.3" # for test purpose only, not really used
psutil = "^5.6"

[tool.poetry.scripts]
akimous = 'akimous.__main__:start'
